Removed internal global barrier in SendRecv
Instead of issuing a shmem_barrier_all, the reception of the message is
confirmed using a flag.
The macros shmem_reset_flag , shmem_set_flag and shmem_wait_flag,
currently defined in domdec_network.c for simplicity,
enables the creation of a non-global barrier to which only
the processes involved on the SendRecv operation will wait, and only if there is a reception involved.
Also, macros/functions to create/remove shared memory have been added to smalloc.
Temporary code to allow a global shared memory buffer for communication has been added
but it is currently disabled (by the means of the non-defined shmem_temporary_buffer variable)