Implemented real and rvec SendRecv operations
The real and rvec variants of the dd_sendrecv routine have been implemented using shmem
by following the same approach used to the dd_sendrecv_int (i.e global reduction if
new temporary buffer is bigger than current , and flag to set if written or not).

A macro have been added to facilitate writing calls to the renew_buf routine,
in the same spirit than the others calls in GROMACs.

Tested with 2, 4 and 8 proc.