Replacing mem. alloc. by shmem equivalents
In this commit all the variables used in the dd_sendrecv2_rvec  routine
are allocated on the symmetric heap. This required changes in many different
files ensuring all PEs allocate memory at the same time.
The variables used in the dd_sendrecv2_rvec routines are:

- state_local: The local state of each PE during the decomposition.
  Notice the global state is not distributed/communicated thus its
  allocation cannot be symmetric (or the program hangs in the barrier).
  Therefore, there are now two routines: the original init_state and
  a shmem specific, init_state_shmem, used during partitioning and state
  redistribution.
  From within the state, all X, V, sd_X and cg_p are allocated in shmem

- f : created in do_md

The function dd_realloc_state_shmem  handles the reallocation of the states
using the shmem equivalent. It uses the maximum of the nalloc across all PEs
to ensure symmetry whenever it is reallocated.