SendRecv2_rvec implementation w/o intermediate buffer
A new routine dd_sendrecv2_rvec_off has been implemented, where
instead of receiving just the address of the memory position to transfer,
the routine receives the base address of the buffer plus an offset.
This offset is communicated to the processor that gets the data so it
reads the data properly.
Otherwise the offset read by the other processor would be incorrect
(bear in mind memory is symmetric).