Replaced vbuf allocation with shrealloc
In order to use nonbuffered routines the calls to realloc
comm->vbuff need to be symmetric.
It is important to notice that the setup_dd_communication has
two modes of operation, in place or not in place.
These modes are not symmetric (i.e a PE can be operating in-place
whilst the rest is not in place).
Originally vbuf2 was ony reallocated if the operating mode was in-place,
thus creating an asymmetric of allocations and deadlocking the code.
Now the reallocation of vbuf2 (and potentially buf_int2 too) is always done.

A potential improvement may be to broadcast the value of InPlace to reallocate only in the
event if one PE is inPlace and not if none of them are operating in such a mode.