pmeredist_x_q implemented w/o MPI Barriers
dd_pmeredist_x_q is now imlemented solely using shmem
and without barriers in the group.
Synchronization is achieved using a counter of data to be read.
Each PE that reads data decrements the counter.
The PE that put the data on the send buffer can only exit the
redistribution when all its data has been read (i.e counter reaches 0).

Tested with 8p and adh.