Non-barrier version of the pme part of the code
Using a counter added to the atc data structure, we can ensure
that PEs only read values of the atom->acum_count that correspond to their current
iteration.
This is guaranteed due to the wait at the end of the dd_pmeredist_x_q.
If other routines in the PME part are ported to shmem and require the use of either a remote count or a remote acum_count, the same mechanism needs to be used.

Tested with 8p and adh