cg_p and sd_X removed from SHMEM in all code paths
Since the routines involving state->cg_p and state->sd_X are no longer implemented in SHMEM
(they are not in the critical path AFAIK), we can reduce the memory footprint of the code
by not putting them on SHMEM at all.
This may alleviate problems under some circumstances.

Tested with 8p and adh