Shared SHMEM control structure
The shmem control structure is now shared by all parts of the code.
The structure is initialised if the code runs in parallel and assigned
to a pointer in the CR structure.
The PME buffers has been moved to this shared control structure,
and are initialised by all PEs.

Some other minor changes to cleanup code.

Tested with 8p, adh.