PME: Fixed bugs with more than 256 cores
The shmem_getmem routine produced weird problems when going
beyond 256 cores, it is replaced by a shmem_float_get in this commit.
Also, the size of the count vectors is now symmetric across all ranks,
by setting the size always to the number of PEs rather than using nslab.
This enables the usage of 2D decomposition in the PME computation.